+++ components/DonationForm/SelectBirthday.tsx
@@ -37,7 +37,6 @@ const useStyles = makeStyles(() =>
 );
 
 
-
 export const SelectBirthday = (props: any) => {
 
   const classes = useStyles();
